adapt OrderBookWidget
